+++ mobile/android/android-components/components/feature/addons/src/test/java/mozilla/components/feature/addons/ui/ExtensionsTest.kt
@@ -6,12 +6,22 @@ package mozilla.components.feature.addons.amo.mozilla.components.feature.addons.
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import mozilla.components.feature.addons.Addon
+import mozilla.components.feature.addons.R
 import mozilla.components.feature.addons.ui.getFormattedAmount
+import mozilla.components.feature.addons.ui.toLocalizedString
 import mozilla.components.feature.addons.ui.translate
 import mozilla.components.feature.addons.ui.translatedName
+import mozilla.components.feature.addons.update.AddonUpdater
+import mozilla.components.feature.addons.update.AddonUpdater.Status.NotInstalled
+import mozilla.components.feature.addons.update.AddonUpdater.Status.SuccessfullyUpdated
+import mozilla.components.feature.addons.update.AddonUpdater.Status.Error
+import mozilla.components.feature.addons.update.AddonUpdater.Status.NoUpdateAvailable
+import mozilla.components.support.test.robolectric.testContext
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
+import java.util.Date
 import java.util.Locale
 
 @RunWith(AndroidJUnit4::class)
@@ -74,5 +84,26 @@ class ExtensionsTest {
 
         Locale.setDefault(Locale("es"))
         assertEquals("1.000", getFormattedAmount(amount))
+    }
+
+    @Test
+    fun toLocalizedString() {
+        var request = AddonUpdater.UpdateAttempt("addonId", Date(), SuccessfullyUpdated)
+        var string = testContext.getString(R.string.mozac_feature_addons_updater_status_successfully_updated)
+
+        assertEquals(string, request.status.toLocalizedString(testContext))
+
+        string = testContext.getString(R.string.mozac_feature_addons_updater_status_no_update_available)
+        request = request.copy(status = NoUpdateAvailable)
+
+        assertEquals(string, request.status.toLocalizedString(testContext))
+
+        string = testContext.getString(R.string.mozac_feature_addons_updater_status_error)
+        request = request.copy(status = Error("error", Exception()))
+
+        assertTrue(request.status?.toLocalizedString(testContext)!!.contains(string))
+
+        request = request.copy(status = NotInstalled)
+        assertEquals("", request.status.toLocalizedString(testContext))
     }
 }